updated dfs
